cffdump: better handling of "draw state"
a6xx seems to get a bit more clever, setting state-groups that are
applicable for binning as well as render pass at the same time.  So we
actually do need to defer evaluating the state groups until draw time,
and then take into consideration the current mode.

Not sure yet if the "CP_SET_MODE" packet is the best way to
differentiate which group-id's are evaluated at draw time, but it
seems like a reasonable way to differentiate BINNING vs DRAW so we'll
use that for now.